backend/advisory-service/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
python-dotenv
redis
